fix: use f64 for optional param ranges, fixes MAMA validation
The range field was (i32, i32) which truncated float bounds like
[0.01, 0.99] to (0, 0), causing `x < 0 || x > 0` which rejects
everything. Now uses (f64, f64) so real-param ranges are preserved.

MAMA's optInFastLimit now correctly validates against [0.01, 0.99].

+++ tools/ta_codegen/src/ir.rs
@@ -31,7 +31,7 @@ pub struct Input {
 pub struct OptInput {
     pub name: String,
     pub param_type: ParamType,
-    pub range: Option<(i32, i32)>,
+    pub range: Option<(f64, f64)>,
     pub default: Option<f64>,
     pub display_name: Option<String>,
     pub hint: Option<String>,
